+++ Planarian.Web/src/Modules/Project/Services/ProjectService.ts
@@ -2,7 +2,7 @@ import { HttpClient } from "../../..";
 import { CreateOrEditProject } from "../Models/CreateOrEditProject";
 import { ProjectVm } from "../Models/ProjectVm";
 import { SelectListItem } from "../../../Shared/Models/SelectListItem";
-import { InviteMember } from "../../../Shared/Components/InviteMember";
+import { InviteMember } from "../../../Shared/Models/InviteMember";
 import { TripVm } from "../../Trips/Models/TripVm";
 
 const baseUrl = "api/projects";
