+++ Planarian.Web/src/Modules/Therion/Models/Coordinates.ts
@@ -1,0 +1,9 @@
+export class Coordinates {
+  x: number;
+  y: number;
+
+  constructor(x: number, y: number) {
+    this.x = x;
+    this.y = y;
+  }
+}
