+++ Planarian.Web/src/Modules/Therion/Models/Station.ts
@@ -1,0 +1,9 @@
+import { Coordinates } from "./Coordinates";
+
+export class Station {
+  constructor(coordinates: Coordinates) {
+    this.coordinates = coordinates;
+  }
+  coordinates: Coordinates;
+  name?: string;
+}
