+++ Planarian.Web/src/Modules/Trips/Components/TripCreateButtonComponent.tsx
@@ -12,9 +12,9 @@ import TextArea from "antd/lib/input/TextArea";
 import { DefaultOptionType } from "antd/lib/select";
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { PropertyLength } from "../../../Shared/Constants";
-import { SettingsService } from "../../Settings/Services/settings.service";
-import { ProjectService } from "../../Project/Services/project.service";
+import { PropertyLength } from "../../../Shared/Constants/PropertyLengthConstant";
+import { SettingsService } from "../../Settings/Services/SettingsService";
+import { ProjectService } from "../../Project/Services/ProjectService";
 import { CreateOrEditTripVm } from "../Models/CreateOrEditTripVm";
 import { TripService } from "../Services/TripService";
 import { nameof } from "../../../Shared/Helpers/StringHelpers";
