+++ Planarian.Web/src/Modules/User/Models/ResetPasswordEmailVm.ts
@@ -1,0 +1,3 @@
+export interface ResetPasswordEmailVm {
+  email: string;
+}
