+++ Planarian.Web/src/Modules/User/Models/UpdatePasswordVm.ts
@@ -2,7 +2,3 @@ export interface UpdatePasswordVm {
   password: string;
   confirmPassword: string;
 }
-
-export interface ResetPasswordEmailVm {
-  email: string;
-}
