+++ Planarian.Web/src/index.tsx
@@ -5,7 +5,7 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import axios from "axios";
 import { message } from "antd";
-import { AuthenticationService } from "./Modules/Authentication/Services/authentication.service";
+import { AuthenticationService } from "./Modules/Authentication/Services/AuthenticationService";
 import { isNullOrWhiteSpace } from "./Shared/Helpers/StringHelpers";
 
 const root = ReactDOM.createRoot(
